+++ c7/22/library/include/KeyedFilePersistence.h
@@ -27,7 +27,7 @@ public:
     {
     }
 
-    unsigned int Size() const 
+	size_t Size() const
     { 
         return 0; 
     }
